Extract AbstractRestApiListingTest from RestApiListingsTest
