+++ testsuite/src/test/java/org/jboss/da/test/client/rest/listings/RestApiListingsTest.java
@@ -2,118 +2,16 @@ package org.jboss.da.test.client.rest.listings;
 
 import static org.apache.commons.io.FileUtils.readFileToString;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.fail;
-
-import org.apache.commons.io.FileUtils;
-import org.jboss.da.test.client.AbstractRestReportsTest;
 import org.jboss.resteasy.client.ClientRequest;
 import org.jboss.resteasy.client.ClientResponse;
-import org.jboss.resteasy.util.GenericType;
-import org.junit.After;
 import org.junit.Test;
-
-import javax.ws.rs.core.MediaType;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-public class RestApiListingsTest extends AbstractRestReportsTest {
+public class RestApiListingsTest extends AbstractRestApiListingTest {
 
     private RequestGenerator generator = new RequestGenerator();
-
-    private enum ListEntityType {
-        BLACK, WHITE, PRODUCT;
-    }
-
-    private enum OperationType {
-        POST, DELETE, PUT;
-    }
-
-    private static final String ENCODING = "utf-8";
-
-    private static final String PATH_FILES_LISTINGS_GAV = "/listings";
-
-    private static final String PATH_WHITE_LIST = "/listings/whitelist";
-
-    private static final String PATH_BLACK_LIST = "/listings/blacklist";
-
-    private static final String PATH_WHITE_LISTINGS_GAV = "/listings/whitelist/gav";
-
-    private static final String PATH_WHITE_ARTIFACTS = "/listings/whitelist/artifacts/gav";
-
-    private static final String PATH_BLACK_LISTINGS_GAV = "/listings/blacklist/gav";
-
-    private static final String PATH_PRODUCT = "/listings/whitelist/product";
-
-    private static final String PATH_PRODUCTS = "/listings/whitelist/products";
-
-    private static final ObjectMapper mapper = new ObjectMapper();
-
-    @After
-    public void dropTables() throws Exception {
-        List<RestWhiteArtifact> whitelistedArtifacts = getAllWhiteArtifactsFromList(PATH_WHITE_LIST);
-        List<RestArtifact> artifacts = whiteToRestArtifactList(whitelistedArtifacts);
-        artifacts.forEach(gav -> removeGavFromList(PATH_WHITE_LISTINGS_GAV, gav));
-
-        List<RestArtifact> blacklistedArtifacts = getAllArtifactsFromList(PATH_BLACK_LIST);
-        blacklistedArtifacts.forEach(gav -> removeGavFromList(PATH_BLACK_LISTINGS_GAV, gav));
-
-        List<RestProduct> products = getAllProductsFromList(PATH_PRODUCTS);
-        products.forEach(product -> removeProductFromList(PATH_PRODUCT, product));
-    }
-
-    private void removeGavFromList(String listUrl, RestArtifact gav) {
-        try {
-            ClientRequest request = createClientRequest(listUrl, mapper.writeValueAsString(gav));
-            request.delete(String.class);
-        } catch (Exception e) {
-            fail("Failed to remove GAV from the list using URL " + listUrl);
-        }
-    }
-
-    private void removeProductFromList(String url, RestProduct product) {
-        try {
-            ClientRequest request = createClientRequest(url, toRestProductRequest(product));
-            request.delete(String.class);
-        } catch (Exception e) {
-            fail("Failed to remove product from the list using URL " + url);
-        }
-
-    }
-
-    private String toRestProductRequest(RestProduct p) {
-        return "{" + "\"name\":" + "\"" + p.getName() + "\"," + "\"version\":" + "\""
-                + p.getVersion() + "\"" + "}";
-    }
-
-    private List<RestArtifact> getAllArtifactsFromList(String listUrl) throws Exception {
-        return processGetRequest(new GenericType<List<RestArtifact>>() {}, restApiURL + listUrl);
-    }
-
-    private List<RestWhiteArtifact> getAllWhiteArtifactsFromList(String listUrl) throws Exception {
-        return processGetRequest(new GenericType<List<RestWhiteArtifact>>() {}, restApiURL
-                + listUrl);
-    }
-
-    private List<RestProduct> getAllProductsFromList(String listUrl) throws Exception {
-        return processGetRequest(new GenericType<List<RestProduct>>() {}, restApiURL + listUrl);
-    }
-
-    private <T> T processGetRequest(GenericType<T> type, String url) throws Exception {
-        ClientRequest request = new ClientRequest(url);
-        request.accept(MediaType.APPLICATION_JSON);
-
-        ClientResponse<T> response = request.get(type);
-
-        if (response.getStatus() != 200)
-            fail("Failed to get entity via REST API");
-
-        return response.getEntity();
-    }
 
     @Test
     public void testAddProduct() throws Exception {
@@ -136,7 +34,7 @@ public class RestApiListingsTest extends AbstractRestReportsTest {
 
         type = "productChangeStatus";
 
-        response = manipulateEntity(ListEntityType.PRODUCT, OperationType.PUT, type, true);
+        manipulateEntity(ListEntityType.PRODUCT, OperationType.PUT, type, true);
     }
 
     @Test
@@ -277,18 +175,18 @@ public class RestApiListingsTest extends AbstractRestReportsTest {
     public void testAddWhitelistedArtifactToBlacklist() throws Exception {
         String type = "productAdd";
 
-        ClientResponse<String> response = manipulateEntity(ListEntityType.PRODUCT,
-                OperationType.POST, type, true);
+        manipulateEntity(ListEntityType.PRODUCT, OperationType.POST, type, true);
 
         // Add artifact to whitelist
         type = generator.returnWhiteArtifactString("org.jboss.da", "dependency-analyzer", "0.3.0",
                 getIdOfProduct("test", "1.0.0"));
 
-        response = manipulateEntity(ListEntityType.WHITE, OperationType.POST, type, true);
+        manipulateEntity(ListEntityType.WHITE, OperationType.POST, type, true);
         // Add artifact to blacklist
         type = "gav";
 
-        response = manipulateEntity(ListEntityType.BLACK, OperationType.POST, type, true);
+        ClientResponse<String> response = manipulateEntity(ListEntityType.BLACK,
+                OperationType.POST, type, true);
 
         checkExpectedResponse(response, "successMessage");
 
@@ -299,25 +197,25 @@ public class RestApiListingsTest extends AbstractRestReportsTest {
     public void testAddMultipleTimeWhitelistedArtifactToBlacklist() throws Exception {
         String type = "productAdd";
 
-        ClientResponse<String> response = manipulateEntity(ListEntityType.PRODUCT,
-                OperationType.POST, type, true);
+        manipulateEntity(ListEntityType.PRODUCT, OperationType.POST, type, true);
 
         type = "productAdd2";
 
-        response = manipulateEntity(ListEntityType.PRODUCT, OperationType.POST, type, true);
+        manipulateEntity(ListEntityType.PRODUCT, OperationType.POST, type, true);
         // Add artifacts to whitelist
         type = generator.returnWhiteArtifactString("org.jboss.da", "dependency-analyzer", "0.3.0",
                 getIdOfProduct("test", "1.0.0"));
 
-        response = manipulateEntity(ListEntityType.WHITE, OperationType.POST, type, true);
+        manipulateEntity(ListEntityType.WHITE, OperationType.POST, type, true);
 
         type = generator.returnWhiteArtifactString("org.jboss.da", "dependency-analyzer", "0.3.0",
                 getIdOfProduct("test", "2.0.0"));
 
-        response = manipulateEntity(ListEntityType.WHITE, OperationType.POST, type, true);
+        manipulateEntity(ListEntityType.WHITE, OperationType.POST, type, true);
         // Add artifact to blacklist
         type = "gav";
-        response = manipulateEntity(ListEntityType.BLACK, OperationType.POST, type, true);
+        ClientResponse<String> response = manipulateEntity(ListEntityType.BLACK,
+                OperationType.POST, type, true);
 
         checkExpectedResponse(response, "successMessage");
     }
@@ -326,16 +224,16 @@ public class RestApiListingsTest extends AbstractRestReportsTest {
     public void testAlreadyAddedWhiteArtifact() throws Exception {
         String type = "productAdd";
 
-        ClientResponse<String> response = manipulateEntity(ListEntityType.PRODUCT,
-                OperationType.POST, type, true);
+        manipulateEntity(ListEntityType.PRODUCT, OperationType.POST, type, true);
 
         type = generator.returnWhiteArtifactString("org.jboss.da", "dependency-analyzer", "0.3.0",
                 getIdOfProduct("test", "1.0.0"));
 
-        response = manipulateEntity(ListEntityType.WHITE, OperationType.POST, type, true);
+        manipulateEntity(ListEntityType.WHITE, OperationType.POST, type, true);
 
         // add second white artifact
-        response = manipulateEntity(ListEntityType.WHITE, OperationType.POST, type, true);
+        ClientResponse<String> response = manipulateEntity(ListEntityType.WHITE,
+                OperationType.POST, type, true);
 
         checkExpectedResponse(response, "successFalse");
     }
@@ -345,25 +243,25 @@ public class RestApiListingsTest extends AbstractRestReportsTest {
         // Add artifacts to whitelist
         String type = "productAdd";
 
-        ClientResponse<String> response = manipulateEntity(ListEntityType.PRODUCT,
-                OperationType.POST, type, true);
+        manipulateEntity(ListEntityType.PRODUCT, OperationType.POST, type, true);
 
         type = "productAdd2";
 
-        response = manipulateEntity(ListEntityType.PRODUCT, OperationType.POST, type, true);
+        manipulateEntity(ListEntityType.PRODUCT, OperationType.POST, type, true);
         // Add artifacts to whitelist
         type = generator.returnWhiteArtifactString("org.jboss.da", "dependency-analyzer", "0.3.0",
                 getIdOfProduct("test", "1.0.0"));
 
-        response = manipulateEntity(ListEntityType.WHITE, OperationType.POST, type, true);
+        manipulateEntity(ListEntityType.WHITE, OperationType.POST, type, true);
 
         type = generator.returnWhiteArtifactString("org.jboss.da", "dependency-analyzer", "0.3.0",
                 getIdOfProduct("test", "2.0.0"));
 
-        response = manipulateEntity(ListEntityType.WHITE, OperationType.POST, type, true);
+        manipulateEntity(ListEntityType.WHITE, OperationType.POST, type, true);
         // Get list
 
-        response = new ClientRequest(restApiURL + PATH_WHITE_LIST).get(String.class);
+        ClientResponse<String> response = new ClientRequest(restApiURL + PATH_WHITE_LIST)
+                .get(String.class);
         assertEquals(200, response.getStatus());
 
         checkExpectedResponse(response, "gavWhiteList");
@@ -440,15 +338,14 @@ public class RestApiListingsTest extends AbstractRestReportsTest {
     public void testCheckRHWhiteArtifact() throws Exception {
         String type = "productAdd";
 
-        ClientResponse<String> response = manipulateEntity(ListEntityType.PRODUCT,
-                OperationType.POST, type, true);
+        manipulateEntity(ListEntityType.PRODUCT, OperationType.POST, type, true);
 
         type = generator.returnWhiteArtifactString("org.jboss.da", "dependency-analyzer",
                 "0.3.0.redhat-1", getIdOfProduct("test", "1.0.0"));
 
-        response = manipulateEntity(ListEntityType.WHITE, OperationType.POST, type, true);
+        manipulateEntity(ListEntityType.WHITE, OperationType.POST, type, true);
 
-        response = new ClientRequest(restApiURL + PATH_WHITE_LIST
+        ClientResponse<String> response = new ClientRequest(restApiURL + PATH_WHITE_LIST
                 + "?groupid=org.jboss.da&artifactid=dependency-analyzer&version=0.3.0.redhat-1")
                 .get(String.class);
         assertEquals(200, response.getStatus());
@@ -465,94 +362,19 @@ public class RestApiListingsTest extends AbstractRestReportsTest {
     public void testCheckNonRHWhiteArtifact() throws Exception {
         String type = "productAdd";
 
-        ClientResponse<String> response = manipulateEntity(ListEntityType.PRODUCT,
-                OperationType.POST, type, true);
+        manipulateEntity(ListEntityType.PRODUCT, OperationType.POST, type, true);
 
         type = generator.returnWhiteArtifactString("org.jboss.da", "dependency-analyzer",
                 "0.3.0.redhat-1", getIdOfProduct("test", "1.0.0"));
 
-        response = manipulateEntity(ListEntityType.WHITE, OperationType.POST, type, true);
+        manipulateEntity(ListEntityType.WHITE, OperationType.POST, type, true);
 
-        response = new ClientRequest(restApiURL + PATH_WHITE_LIST
+        ClientResponse<String> response = new ClientRequest(restApiURL + PATH_WHITE_LIST
                 + "?groupid=org.jboss.da&artifactid=dependency-analyzer&version=0.3.0")
                 .get(String.class);
         assertEquals(200, response.getStatus());
 
         checkExpectedResponse(response, "gavRhResponse");
-    }
-
-    private ClientResponse<String> manipulateEntity(ListEntityType entity, OperationType operation,
-            String file, Boolean checkSuccess) throws Exception {
-        String type = file;
-        String requestString = null;
-        if (entity.equals(ListEntityType.WHITE) && operation.equals(OperationType.POST)) {
-            requestString = file;
-        } else {
-            File jsonRequestFile = getJsonRequestFile(PATH_FILES_LISTINGS_GAV, type);
-            requestString = FileUtils.readFileToString(jsonRequestFile, ENCODING);
-        }
-        String path = null;
-        switch (entity) {
-            case WHITE:
-                path = PATH_WHITE_LISTINGS_GAV;
-                break;
-            case BLACK:
-                path = PATH_BLACK_LISTINGS_GAV;
-                break;
-            case PRODUCT:
-                path = PATH_PRODUCT;
-                break;
-        }
-
-        ClientRequest request = createClientRequest(path, requestString);
-        ClientResponse<String> response = null;
-        switch (operation) {
-            case POST:
-                response = request.post(String.class);
-                break;
-
-            case DELETE:
-                response = request.delete(String.class);
-                break;
-            case PUT:
-                response = request.put(String.class);
-                break;
-        }
-        if (checkSuccess)
-            assertEquals(200, response.getStatus());
-        return response;
-    }
-
-    private void checkBlacklistingArtifactTest(String artifact1, String artifact2, String addToBL)
-            throws Exception {
-        manipulateEntity(ListEntityType.WHITE, OperationType.POST, artifact1, true);
-        if (artifact2 != null) {
-            manipulateEntity(ListEntityType.WHITE, OperationType.POST, artifact2, true);
-        }
-        manipulateEntity(ListEntityType.BLACK, OperationType.POST, addToBL, true);
-
-        assertEquals(0, getAllArtifactsFromList(PATH_WHITE_LIST).size());
-        assertEquals(1, getAllArtifactsFromList(PATH_BLACK_LIST).size());
-        assertEquals("1.0.0", getAllArtifactsFromList(PATH_BLACK_LIST).get(0).getVersion());
-    }
-
-    private void checkArtifactTest(String artifact1, String artifact2, String testVersion,
-            String expectedResult) throws Exception {
-        manipulateEntity(ListEntityType.WHITE, OperationType.POST, artifact1, true);
-
-        if (artifact2 != null) {
-            manipulateEntity(ListEntityType.WHITE, OperationType.POST, artifact2, true);
-        }
-
-        System.out.println(restApiURL + PATH_WHITE_ARTIFACTS
-                + "?groupid=org.jboss.da&artifactid=dependency-analyzer&version=" + testVersion);
-
-        ClientResponse<String> response = new ClientRequest(restApiURL + PATH_WHITE_ARTIFACTS
-                + "?groupid=org.jboss.da&artifactid=dependency-analyzer&version=" + testVersion)
-                .get(String.class);
-        System.out.println("RESULT:" + response.getEntity());
-        assertEquals(200, response.getStatus());
-        checkExpectedResponse(response, expectedResult);
     }
 
     private void checkExpectedResponse(ClientResponse<String> response, String expectedFile)
@@ -561,21 +383,4 @@ public class RestApiListingsTest extends AbstractRestReportsTest {
         assertEqualsJson(readFileToString(expectedResponseFile), response.getEntity(String.class));
     }
 
-    private List<RestArtifact> whiteToRestArtifactList(List<RestWhiteArtifact> whites) {
-        List<RestArtifact> artifacts = new ArrayList<RestArtifact>();
-        for (RestWhiteArtifact w : whites) {
-            RestArtifact a = new RestArtifact();
-            a.setArtifactId(w.getGav().artifactId);
-            a.setGroupId(w.getGav().getGroupId());
-            a.setVersion(w.getGav().getVersion());
-            artifacts.add(a);
-        }
-        return artifacts;
-    }
-
-    private long getIdOfProduct(String name, String version) throws Exception {
-        ClientResponse<RestProduct[]> r = new ClientRequest(restApiURL + PATH_PRODUCT + "?name="
-                + name + "&version=" + version).get(RestProduct[].class);
-        return r.getEntity()[0].getId();
-    }
 }
